% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binge_params.R
\docType{data}
\name{binge_params_stapm_scot}
\alias{binge_params_stapm_scot}
\title{Parameters to estimate amount drunk on single occasions - Scotland - STAPM version}
\format{
A list of four data tables (1 = Negative binomial regression model for the number of weekly drinking occasions, 
2 = Fitted Heckman selection model for probability that an individual drinks on at least 3 separate occasions during the diary period, 
3 = Fitted Heckman outcome regression results for the standard deviation in the quantity of alcohol consumed in a drinking occasion, 
4 = average height and weight)
}
\source{
Hill-McManus et al 2014. "Estimation of usual occasion-based individual drinking patterns using diary survey data". https://doi.org/https://doi.org/10.1016/j.drugalcdep.2013.09.022.
}
\usage{
binge_params_stapm_scot
}
\description{
Scottish version of `binge_params_stapm` based on the Scottish Health Survey years 2008 to 2019.
}
\keyword{datasets}
