% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_names.R
\docType{data}
\name{disease_groups}
\alias{disease_groups}
\title{Groupings of diseases into disease types}
\format{
A data.table containing disease groupings following the schemes used in 
the Royal College of Physicians report 'Hiding in Plain Sight' Chapter 3 \insertCite{RCP2018}{tobalcepi} 
and \insertCite{Angus2018;textual}{tobalcepi}. 
Also includes a column of formatted disease names and an indication of which diseases are related to tobacco only, alcohol only or both
}
\usage{
disease_groups
}
\description{
Groupings of diseases into disease types
}
\references{
\insertRef{RCP2018}{tobalcepi}    
\insertRef{Angus2018}{tobalcepi}
}
\keyword{datasets}
