% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_match.R
\name{\%fin\%}
\alias{\%fin\%}
\title{Fast matching}
\usage{
x \%fin\% table
}
\arguments{
\item{x}{values to be matched}

\item{table}{values to be matched against}
}
\value{
Logical indicating match or not
}
\description{
A faster version of `%in%`.
}
\details{
See [fastmatch::fmatch()]
}
\examples{

\dontrun{

x \%fin\% y

}



}
\keyword{internal}
