% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hse_data_smoking.R
\docType{data}
\name{hse_data_smoking}
\alias{hse_data_smoking}
\title{Health Survey for England data used for modelling smoking}
\format{A data table}
\source{
see the processing code in the data-raw folder, which uses the hseclean R package
}
\usage{
hse_data_smoking
}
\description{
For years 2001-2016.
}
\keyword{datasets}
