% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAFcalc.R
\name{PAFcalc}
\alias{PAFcalc}
\title{Calculate Population Attributable Fractions \lifecycle{maturing}}
\usage{
PAFcalc(
  data,
  substance,
  tob_include_risk_in_former_smokers = TRUE,
  use_weights = FALSE,
  year_range = "all",
  pool = FALSE,
  subgroups = c("sex", "age_cat")
)
}
\arguments{
\item{data}{Data table of individual characteristics}

\item{substance}{Whether to compute relative risks for just alcohol ("alc"),
just tobacco ("tob") or joint risks for tobacco and alcohol ("tobalc").}

\item{tob_include_risk_in_former_smokers}{Logical - whether the residual risks of smoking in former smokers
should be considered (defaults to TRUE).}

\item{use_weights}{Logical - should the calculation account for survey weights. Defaults to FALSE.
Weight variable must be called "wt_int".}

\item{year_range}{Either an integer vector of the years to be selected or "all". Defaults to "all".}

\item{pool}{Logical - should the years selected be pooled. Defaults to FALSE.}

\item{subgroups}{Character vector - the variable names of the subgroups used to stratify the estimates.}
}
\value{
Returns a data.table containing the estimated PAFs.
}
\description{
Uses \code{RRFunc()} and \code{subgroupRisk()} to 
calculate population attributable fractions 
based on the survey data provided.
}
\examples{
\dontrun{

tobacco_pafs <- PAFcalc(
 data = test_data,
 substance = "tob",
 tob_include_risk_in_former_smokers = TRUE,
 use_weights = TRUE,
 year_range = "all",
 pool = TRUE,
 subgroups = c("sex", "age_cat", "imd_quintile")
)

}

}
