% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tob_alc_risk_int.R
\docType{data}
\name{tob_alc_risk_int}
\alias{tob_alc_risk_int}
\title{Synergistic effects of tobacco and alcohol risks}
\format{
A data table. We currently include estimates of synergistic effects for 
oral, pharyngeal, laryngeal and oesophageal cancers 
(we describe these data and their sources in our tobacco risk functions report 
\insertCite{webster2018risk;textual}{tobalcepi}). 
The data sources we use are \insertCite{Prabhu2014;textual}{tobalcepi} 
 and \insertCite{Hashibe2009;textual}{tobalcepi}.
}
\usage{
tob_alc_risk_int
}
\description{
Synergistic effects of tobacco and alcohol risks
}
\references{
\insertRef{webster2018risk}{tobalcepi}  

\insertRef{Hashibe2009}{tobalcepi}  

\insertRef{Prabhu2014}{tobalcepi}
}
\keyword{datasets}
