% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_names.R
\docType{data}
\name{tob_disease_names}
\alias{tob_disease_names}
\title{Names of tobacco-related diseases}
\format{
A character vector containing the names of the 52 disease categories 
of smoking related diseases \insertCite{webster2018risk;textual}{tobalcepi}.
}
\usage{
tob_disease_names
}
\description{
Names of tobacco-related diseases
}
\references{
\insertRef{webster2018risk}{tobalcepi}
}
\keyword{datasets}
