% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobacco_lag_times.R
\docType{data}
\name{tobacco_lag_times}
\alias{tobacco_lag_times}
\title{Tobacco lag times}
\format{
A data table
}
\source{

}
\usage{
tobacco_lag_times
}
\description{
Tobacco lag times
}
\keyword{datasets}
