% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobacco_relative_risks.R
\docType{data}
\name{tobacco_relative_risks}
\alias{tobacco_relative_risks}
\title{Tobacco relative risks}
\format{
A data table
}
\source{

}
\usage{
tobacco_relative_risks
}
\description{
Tobacco relative risks
}
\keyword{datasets}
