% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TobAlcInt.R
\name{TobAlcInt}
\alias{TobAlcInt}
\title{Risk interaction between tobacco and alcohol}
\usage{
TobAlcInt(data, disease = "Pharynx", alcohol_var = "weekmean",
  tobacco_var = "smk.state", rr_data, account_for_synergy = TRUE)
}
\arguments{
\item{data}{Data table}

\item{disease}{Character}

\item{alcohol_var}{Character}

\item{tobacco_var}{Character}

\item{rr_data}{Data table}

\item{account_for_synergy}{Logical}
}
\value{
Returns a numeric vector containing  of each individual's relative risks for the tobacco-related disease
specified by "disease".
}
\description{
Assigns the disease-specific interaction term (synergy index) appropriate to each
individual's tobacco and alcohol consumption.
}
