% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alc_disease_names.R
\docType{data}
\name{alc_disease_names}
\alias{alc_disease_names}
\title{Names of alcohol-related diseases}
\format{
A data table
}
\source{

}
\usage{
alc_disease_names
}
\description{
Names of alcohol-related diseases
}
\keyword{datasets}
