% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tob_disease_names.R
\docType{data}
\name{disease_groups}
\alias{disease_groups}
\title{Groupings of smoking-related diseases into disease types}
\format{
A data.table containing disease groupings following the scheme used in 
the Royal College of Physicians report 'Hiding in Plain Sight' Chapter 3 \insertCite{RCP2018}{tobalcepi}. 
Also includes a column of formatted disease names.
}
\usage{
disease_groups
}
\description{
Groupings of smoking-related diseases into disease types
}
\references{
\insertRef{RCP2018}{tobalcepi}
}
\keyword{datasets}
