% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_groups.R
\docType{data}
\name{disease_groups}
\alias{disease_groups}
\title{Groupings of smoking-related diseases into disease types}
\format{
A data table
}
\source{
Following the scheme used in the Royal College of Physicians report 'Hiding in Plain Signt' chapter 3.
}
\usage{
disease_groups
}
\description{
Groupings of smoking-related diseases into disease types
}
\keyword{datasets}
