% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAFcalc.R
\name{PAFcalc}
\alias{PAFcalc}
\title{Calculate Population Attributable Fractions \lifecycle{maturing}}
\usage{
PAFcalc(
  data,
  substance,
  tob_include_risk_in_former_smokers = TRUE,
  alc_protective = TRUE,
  alc_wholly_chronic_thresholds = c(2, 2),
  alc_wholly_acute_thresholds = c(3, 4),
  grams_ethanol_per_unit = 8,
  use_weights = FALSE,
  year_range = "all",
  pool = FALSE,
  subgroups = c("sex", "age_cat"),
  tobalc_include_int = FALSE
)
}
\arguments{
\item{data}{Data table of individual characteristics}

\item{substance}{Whether to compute relative risks for just alcohol ("alc"),
just tobacco ("tob") or joint risks for tobacco and alcohol ("tobalc").}

\item{tob_include_risk_in_former_smokers}{Logical - whether the residual risks of smoking in former smokers
should be considered (defaults to TRUE).}

\item{alc_protective}{Logical - whether to include the protective effects of
alcohol in the risk function. Defaults to TRUE. If TRUE, then the part of the risk function < 1 is set to equal 1.}

\item{alc_wholly_chronic_thresholds}{Numeric vector - the thresholds in UK standard units of alcohol per day 
over which individuals begin to experience an elevated risk
 for chronic diseases that are wholly attributable to alcohol. Input in the order c(female, male). 
 Defaults to the current UK healthy drinking threshold of 14 units/week for females and males, or 2 units/day.}

\item{alc_wholly_acute_thresholds}{Numeric vector - the thresholds in UK standard units of alcohol /day over
which individuals begin to experience an elevated risk
for acute diseases that are wholly attributable to alcohol. Input in the form c(female, male). 
Defaults to 3 units/day for females and 4 units/day for males.}

\item{grams_ethanol_per_unit}{Numeric value giving the conversion factor for the number of grams of pure
ethanol in one UK standard unit of alcohol.}

\item{use_weights}{Logical - should the calculation account for survey weights. Defaults to FALSE.
Weight variable must be called "wt_int".}

\item{year_range}{Either an integer vector of the years to be selected or "all". Defaults to "all".}

\item{pool}{Logical - should the years selected be pooled. Defaults to FALSE.}

\item{subgroups}{Character vector - the variable names of the subgroups used to stratify the estimates.}

\item{tobalc_include_int}{Logical - in computing joint relative risks for tobacco and alcohol,
should a (synergistic/multiplicative) interaction between exposure to tobacco and alcohol be included.
Defaults to FALSE. If TRUE, then only interactive effects for oesophageal, pharynx, oral cavity and larynx cancers
are considered.}
}
\value{
Returns a data.table containing the estimated PAFs.
}
\description{
Uses \code{RRFunc()} and \code{subgroupRisk()} to 
calculate population attributable fractions 
based on the survey data provided.
}
\examples{
\dontrun{

tobacco_pafs <- PAFcalc(
 data = test_data,
 substance = "tob",
 tob_include_risk_in_former_smokers = TRUE,
 use_weights = TRUE,
 year_range = "all",
 pool = TRUE,
 subgroups = c("sex", "age_cat", "imd_quintile")
)

}

}
