% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tob_alc_risk_int.R
\docType{data}
\name{tob_alc_risk_int}
\alias{tob_alc_risk_int}
\title{Synergstic effects of tobacco and alcohol risks}
\format{
A data table
}
\source{

}
\usage{
tob_alc_risk_int
}
\description{
Synergstic effects of tobacco and alcohol risks
}
\keyword{datasets}
