% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tob_disease_names.R
\docType{data}
\name{tob_disease_names}
\alias{tob_disease_names}
\title{Names of tobacco-related diseases}
\format{
A data table
}
\source{

}
\usage{
tob_disease_names
}
\description{
Names of tobacco-related diseases
}
\keyword{datasets}
