% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobacco_relative_risks.R
\docType{data}
\name{tobacco_relative_risks}
\alias{tobacco_relative_risks}
\title{Tobacco relative risks}
\format{
A data table containing the relative risks of 52 disease categories 
for current vs. never smokers \insertCite{webster2018risk;textual}{tobalcepi}.
}
\usage{
tobacco_relative_risks
}
\description{
Tobacco relative risks
}
\references{
\insertRef{webster2018risk}{tobalcepi}
}
\keyword{datasets}
