% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobalcepi-package.R
\docType{package}
\name{tobalcepi-package}
\alias{tobalcepi}
\alias{tobalcepi-package}
\title{tobalcepi: Risk Functions and Attributable Fractions for Tobacco and Alcohol}
\description{
Functions to assign relative risks of disease to individuals based on their tobacco and/or alcohol consumption, and to estimate the attributable fractions of disease.
}
\author{
\strong{Maintainer}: Duncan Gillespie \email{duncan.gillespie@sheffield.ac.uk} (\href{https://orcid.org/0000-0003-3450-5747}{ORCID})

Authors:
\itemize{
  \item Laura Webster
  \item Maddy Henney
  \item Colin Angus \email{c.r.angus@sheffield.ac.uk}
  \item Alan Brennan \email{a.brennan@sheffield.ac.uk}
}

}
\keyword{internal}
