% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TobAlcInt.R
\name{TobAlcInt}
\alias{TobAlcInt}
\title{Risk interaction between tobacco and alcohol}
\usage{
TobAlcInt(
  data,
  disease = "Pharynx",
  alcohol_var = "weekmean",
  tobacco_var = "smk.state",
  rr.data,
  account_for_synergy = TRUE
)
}
\arguments{
\item{data}{Data table - containing the individual characteristics of smokers and drinkers}

\item{disease}{Character}

\item{alcohol_var}{Character}

\item{tobacco_var}{Character}

\item{rr.data}{Data table}

\item{account_for_synergy}{Logical}
}
\value{
Returns a numeric vector containing  of each individual's relative risks for the tobacco-related disease
specified by "disease".
}
\description{
\lifecycle{maturing}
}
\details{
Assigns the disease-specific interaction term (synergy index) appropriate to each
individual's tobacco and alcohol consumption.

We currently include estimates of synergistic effects for 
oral, pharyngeal, laryngeal and oesophageal cancers 
(we describe these data and their sources in our tobacco risk functions report 
\insertCite{webster2018risk;textual}{tobalcepi}). 
The data sources we use are \insertCite{Prabhu2014;textual}{tobalcepi} 
 and \insertCite{Hashibe2009;textual}{tobalcepi}.
We apply these effects by scaling the joint risks by a 'synergy index', 
which takes the result of a meta-analysis of the additional
risk faced by people because they consume both tobacco and alcohol.
}
\examples{

\dontrun{

TobAlcInt()

}


}
\references{
\insertRef{webster2018risk}{tobalcepi}  

\insertRef{Hashibe2009}{tobalcepi}  

\insertRef{Prabhu2014}{tobalcepi}
}
